Fix prosemirror folder not being mjs type defs, moved ProseMirror def
