+++ src/foundry/prosemirror/extensions.mjs.d.ts
