+++ src/foundry/prosemirror/index.d.ts
@@ -1,13 +1,17 @@
-import './content-link-plugin';
-import './dirty-plugin';
-import './dropdown';
-import './extensions';
-import './image-plugin';
-import './input-rules';
-import './keymaps';
-import './menu';
-import './plugin';
-import './prosemirror';
-import './schema';
-import './string-serializer';
-import './util';
+import './content-link-plugin.mjs';
+import './dirty-plugin.mjs';
+import './dropdown.mjs';
+import './extensions.mjs';
+import './image-plugin.mjs';
+import './input-rules.mjs';
+import './keymaps.mjs';
+import './menu.mjs';
+import './plugin.mjs';
+import * as ProseMirrorObject from './prosemirror.mjs';
+import './schema.mjs';
+import './string-serializer.mjs';
+import './util.mjs';
+
+declare global {
+  let ProseMirror: typeof ProseMirrorObject;
+}
