+++ src/foundry/prosemirror/plugin.mjs.d.ts
