+++ src/foundry/prosemirror/prosemirror.mjs.d.ts
@@ -1,18 +1,18 @@
 import { EditorState, AllSelection, Plugin } from 'prosemirror-state';
 import { EditorView } from 'prosemirror-view';
 import { Schema, DOMParser, DOMSerializer } from 'prosemirror-model';
-import ProseMirrorInputRules from './input-rules';
+import ProseMirrorInputRules from './input-rules.mjs';
 import { keymap } from 'prosemirror-keymap';
-import ProseMirrorKeyMaps from './keymaps';
-import ProseMirrorMenu from './menu';
+import ProseMirrorKeyMaps from './keymaps.mjs';
+import ProseMirrorMenu from './menu.mjs';
 import './extensions';
 import * as collab from 'prosemirror-collab';
 import { Step } from 'prosemirror-transform';
-import { parseHTMLString, serializeHTMLString } from './util';
-import { schema as defaultSchema } from './schema';
-import ProseMirrorImagePlugin from './image-plugin';
-import ProseMirrorDirtyPlugin from './dirty-plugin';
-import ProseMirrorContentLinkPlugin from './content-link-plugin';
+import { parseHTMLString, serializeHTMLString } from './util.mjs';
+import { schema as defaultSchema } from './schema.mjs';
+import ProseMirrorImagePlugin from './image-plugin.mjs';
+import ProseMirrorDirtyPlugin from './dirty-plugin.mjs';
+import ProseMirrorContentLinkPlugin from './content-link-plugin.mjs';
 
 declare const dom: {
   parser: DOMParser;
