+++ src/foundry/prosemirror/schema.mjs.d.ts
