+++ src/foundry/prosemirror/string-serializer.mjs.d.ts
